+++ tests/test_main_window_functions/test_rename_scenario.py
@@ -33,29 +33,40 @@ def test_rename_scenario(qtbot):  # noqa: PLR0915
 
     QtW.QInputDialog = NewQDialog
 
+    main_window.dia.setWindowTitle(main_window.dia.windowTitle()[:-1])
+    main_window.changedFile = False
+    assert main_window.dia.windowTitle()[-1] != "*"
     # get old item name
     old_name = main_window.list_widget_scenario.item(0).text()
     # enter nothing in the text box and not change the name
     qtbot.mouseClick(main_window.button_rename_scenario, QtC.Qt.MouseButton.LeftButton, delay=1)
     # check if the name stays the old one
     assert main_window.list_widget_scenario.item(0).text() == old_name
+    assert main_window.dia.windowTitle()[-1] != "*"
     # just close the dialog window
     qtbot.mouseClick(main_window.button_rename_scenario, QtC.Qt.MouseButton.LeftButton, delay=1)
     # check if the name stays the old one
     assert main_window.list_widget_scenario.item(0).text() == old_name
+    assert main_window.dia.windowTitle()[-1] != "*"
     # abort the dialog window by button
     response = QtW.QMessageBox.Close
     qtbot.mouseClick(main_window.button_rename_scenario, QtC.Qt.MouseButton.LeftButton, delay=1)
     # check if the name stays the old one
     assert main_window.list_widget_scenario.item(0).text() == old_name
+    assert main_window.dia.windowTitle()[-1] != "*"
     # change the name
     response = QtW.QDialog.Accepted
     qtbot.mouseClick(main_window.button_rename_scenario, QtC.Qt.MouseButton.LeftButton, delay=1)
     # check the name has been changed correctly
     assert main_window.list_widget_scenario.item(0).text() == scenario_name
+    assert main_window.dia.windowTitle()[-1] == "*"
     # check if the name can be changed by double click
     ret_scenario_name = scenario_name_2
+    main_window.dia.setWindowTitle(main_window.dia.windowTitle()[:-1])
+    main_window.changedFile = False
+    assert main_window.dia.windowTitle()[-1] != "*"
     main_window.list_widget_scenario.doubleClicked.emit(main_window.list_widget_scenario.model().index(0, 0))
     # check the name has been changed correctly
     assert main_window.list_widget_scenario.item(0).text() == scenario_name_2
+    assert main_window.dia.windowTitle()[-1] == "*"
     close_tests(main_window, qtbot)